# Optimize LazyHTML.Tree.to_html with large trees