@ypconstante 2025-08-07
Benchee script is in the PR description, [html files are the ones used by Floki benchmark](https://github.com/philss/floki/blob/main/benchs/documents.tar.gz).

On main branch - `ERL_COMPILER_OPTIONS=bin_opt_info mix compile --force` 

```
ERL_COMPILER_OPTIONS=bin_opt_info mix compile --force
Compiling 3 files (.ex)
     warning: OPTIMIZED: match context reused
     │
 111 │        do: append_text(rest, text, whitespace_size + 1, ctx, html)
     │        ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
     │
     └─ lib/lazy_html/tree.ex:111

     warning: OPTIMIZED: match context reused
     │
 123 │        do: append_escaped(rest, text, 0, whitespace_size, html)
     │        ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
     │
     └─ lib/lazy_html/tree.ex:123

     warning: OPTIMIZED: match context reused
     │
 164 │       append_escaped(rest, text, offset + size + 1, 0, html)
     │       ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
     │
     └─ lib/lazy_html/tree.ex:164

     warning: OPTIMIZED: match context reused
     │
 164 │       append_escaped(rest, text, offset + size + 1, 0, html)
     │       ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
     │
     └─ lib/lazy_html/tree.ex:164

     warning: OPTIMIZED: match context reused
     │
 164 │       append_escaped(rest, text, offset + size + 1, 0, html)
     │       ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
     │
     └─ lib/lazy_html/tree.ex:164

     warning: OPTIMIZED: match context reused
     │
 164 │       append_escaped(rest, text, offset + size + 1, 0, html)
     │       ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
     │
     └─ lib/lazy_html/tree.ex:164

     warning: OPTIMIZED: match context reused
     │
 164 │       append_escaped(rest, text, offset + size + 1, 0, html)
     │       ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
     │
     └─ lib/lazy_html/tree.ex:164

     warning: OPTIMIZED: match context reused
     │
 169 │     append_escaped(rest, text, offset, size + 1, html)
     │     ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
     │
     └─ lib/lazy_html/tree.ex:169
```