@jonatanklosko 2025-08-07
> In theory the binary should be faster and use fewer memory since we are reusing the match context

Just to clarify, it's not about match context (which has to do with recursive traversal, and that's the same in both implementations), it's about the runtime optimising binary appends. I don't think `bin_opt_info` will tell us any difference, because in the iodata implementation we are simply not building a binary :D