@josevalim 2025-08-07
> Just to clarify, it's not about match context (which has to do with recursive traversal, and that's the same in both implementations), it's about the runtime optimising binary appends

In my mind, "runtime optimizing binary appends" is the match context handling. If the match context was not optimized, then we would not see this optimization. But I found it weird nothing was warned about no match context being created and anothing about the append_attrs or to_html functions.